+++ test/plugins/test_annotations_check.py
@@ -239,3 +239,31 @@ def test_no_duplicate_annotation_errors():
         "2:annotation-missing-token:missing non-optional annotation: '.. setting_name:'"
     }
     assert expected == messages
+
+
+def test_missing_annotation():
+    source = """
+    # .. toggle_name: MYTOGGLE1
+    waffle1 = WaffleFlag('MYTOGGLE1')
+    waffle2 = WaffleFlag('MYTOGGLE2')
+    """
+    messages = run_pylint(source, "toggle-missing-annotation")
+    expected = {
+        "4:toggle-missing-annotation:missing feature toggle annotation"
+    }
+    assert expected == messages
+
+
+def test_missing_annotation_for_unnamed_toggle():
+    source = """
+    # annotated waffle flag
+    # .. toggle_name: MYTOGGLE1
+    waffle1 = CourseWaffleFlag()
+    # unannotated waffle flag
+    waffle2 = ExperimentWaffleFlag()
+    """
+    messages = run_pylint(source, "toggle-missing-annotation")
+    expected = {
+        "6:toggle-missing-annotation:missing feature toggle annotation"
+    }
+    assert expected == messages